Surface errors during searching & show component displaynames in results
An error in one bank while searching was swallowed by the results service,
causing the results component to never receive it.
This in turn caused the results stream to never terminate, and the spinners
to stay indefinitely.

Now when an error occurs during search, it is shown using a warning below the
results table.

Also retrieve a component's displayname when displaying a hit in the results table.