Extract treebank declarations  from config.php
There is now a /treebanks directory where php files declaring a treebank and its components should be placed. Included are some example files for the corpora previously built in to the config example.
Components can now declare the server address they are hosted (referring to the ip/hostname and port to access a possibly external basex server). If not declared, uses the host/ip from the parent treebank.

Also clarify and simplify the code dealing with grinded treebanks.
There are now 3 ways in which a component is mapped to a (or multiple) basex databases to query.

For the basic ungrinded corpus, the name of the component is assumed to be the basex database name. For Vincent's import script this is usually ${corpusName}_ID_${componentName}. So CGN_ID_NA for the NA component in the CGN corpus for example.

For very large corpora (sonar!) a single component may be split up into multiple basex databases. These are defined by the .lst files in /treebank-parts. Every time a component is looked up, if a file exists by name of the component, the databases used are read from that file instead of using the component's name.

For grinded corpora, the database names are the component name followed by a grind category (for example, top--%ww). In the case the xpath is determined to run faster on ungrinded data, the usual process described above is used.

- improve documentation in the code in several places.
- remove "already" from the results responses, instead use remainingComponents and remainingDatabases.
- move gretel-upload's basex host/port declaration out of databaseGroups into its own fields.
- improve validation of groups/variants in treebanks
- simplify show-tree
- simplify treebank-counts (for retrieving total number of results)
- remove global dbuser dbpwd in lieu for defaults in treebanks that don't declare their own (admin, admin)
- fix invalid query generated when requesting context on a grinded corpus
- fix bug in treebank selection for corpora without groups but no variants