Accept full results response in frontend in case of no results
The backend now stops searching after finishing one component or
database and gives all details of the search, so that the frontend can
continue searching. The results interfaces and mapping in frontend has
now been changed so that searching continues until searching all
databases and components has finished.

TODO: not exhaustively tested, and the results limit does not work now -
to be fixed before PR.